+++ shopify/packages/ajax-cart/src/index.ts
@@ -30,16 +30,14 @@ const AjaxCart = mergeProps(Cart.query, {
   addAttributes(input: UpdateAttributes.Input) {
     const mergedAttributes = getNormalizedAttributes();
     return UpdateAttributes.mutation.mutateAsync({
-      ...mergedAttributes.private.record,
-      ...mergedAttributes.public.record,
+      ...mergedAttributes,
       ...input,
     });
   },
   addAttribute(input: { key: string; value: UpdateAttributes.Input[string] }) {
     const mergedAttributes = getNormalizedAttributes();
     return UpdateAttributes.mutation.mutateAsync({
-      ...mergedAttributes.private.record,
-      ...mergedAttributes.public.record,
+      ...mergedAttributes,
       [input.key]: input.value,
     });
   },
@@ -47,16 +45,14 @@ const AjaxCart = mergeProps(Cart.query, {
   removeAttributes(input: string[]) {
     const mergedAttributes = getNormalizedAttributes();
     return UpdateAttributes.mutation.mutateAsync({
-      ...mergedAttributes.private.record,
-      ...mergedAttributes.public.record,
+      ...mergedAttributes,
       ...Object.fromEntries(input.map((key) => [key, null])),
     });
   },
   removeAttribute(key: string) {
     const mergedAttributes = getNormalizedAttributes();
     return UpdateAttributes.mutation.mutateAsync({
-      ...mergedAttributes.private.record,
-      ...mergedAttributes.public.record,
+      ...mergedAttributes,
       [key]: null,
     });
   },