+++ shopify/packages/ajax-cart/src/mutations/replaceItem.ts
@@ -39,8 +39,7 @@ export const mutation = useMutation(
             id: item.id,
             ...tempValues,
             properties: {
-              ...tempValues.properties.private.record,
-              ...tempValues.properties.public.record,
+              ...tempValues.properties,
             },
           },
         ],